split export test into separate files
